Revert "add alternate ramp bigint backend"
This reverts commit 539e6201b82f20438f6f31e312cda767301e68c7.
