+++ Cargo.toml
@@ -4,21 +4,15 @@ version = "0.1.0"
 authors = ["Tom Milligan <tom@reinfer.io>"]
 edition = "2018"
 
-[features]
-backend-num = ["num-bigint"]
-backend-ramp = ["ramp"]
-default = ["backend-num"]
-
 [dependencies]
 cairo-rs = { version = "0.6.0", features = ["png"] }
 clap = "^2.33.0"
 env_logger = "^0.6.1"
 lazy_static = "^1.3.0"
 log = "^0.4.6"
-num-bigint = { version = "^0.2.2", optional = true }
+num-bigint = "^0.2.2"
 num-traits = "^0.2.6"
 num_cpus = "^1.10.0"
 png = "0.14.0"
-ramp = { version = "0.5.1", optional = true }
 threadpool = "^1.7.1"
