+++ Makefile
@@ -1,15 +1,11 @@
-.PHONY: build-ramp dev integrate test
-
-build-ramp:
-	cargo +nightly build --no-default-features --features backend-ramp
+.PHONY: dev integrate test
 
 dev:
 	rustup component add rustfmt
-	rustup toolchain install nightly
 
 integrate:
 	./integrate/check
 
 test:
 	cargo fmt -- --check
-	./test/check
+	cargo test --locked
